requirements.txt
alembic==1.11.1
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6 
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
mako==1.2.4
markupsafe==2.1.3
numpy==1.25.0
pandas==2.0.2
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
schedule==1.2.2
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.17
starlette==0.26.1
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
psycopg2-binary==2.9.6
